visualization_service/requirements.txt
python-dotenv
pandas
supabase
streamlit